pyproject.toml
[tool.poetry]
name = "django-lab823"
version = "0.1.0"
description = ""
authors = ["minheonjun <tamunavollt1@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
pillow = "^10.0.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
